+++ frontend/src/modules/Environments/components/EnvironmentTeamInviteForm.js
@@ -54,8 +54,8 @@ export const EnvironmentTeamInviteForm = (props) => {
         setGroupOptions(
           response.data.listGroups.map((g) => ({
             ...g,
-            value: g.groupName,
-            label: g.groupName
+            value: g.groupUri,
+            label: g.groupUri
           }))
         );
       } else {
