+++ frontend/src/modules/Organizations/components/index.js
@@ -3,4 +3,4 @@ export * from './OrganizationListItem';
 export * from './OrganizationOverview';
 export * from './OrganizationTeamInviteEditForm';
 export * from './OrganizationTeamInviteForm';
-export * from './OrganizationTeams';
+export * from './TeamsTableCard';
