+++ frontend/src/modules/Organizations/services/index.js
@@ -3,6 +3,7 @@ export * from './createOrganization';
 export * from './inviteGroup';
 export * from './listOrganizationEnvironments';
 export * from './listOrganizationGroups';
+export * from './listOrganizationReadOnlyGroups';
 export * from './listOrganizations';
 export * from './removeGroup';
 export * from './updateOrganization';
