+++ frontend/src/modules/Organizations/services/listOrganizationReadOnlyGroups.js
@@ -1,0 +1,31 @@
+import { gql } from 'apollo-boost';
+
+export const listOrganizationReadOnlyGroups = ({
+  filter,
+  organizationUri
+}) => ({
+  variables: {
+    organizationUri,
+    filter
+  },
+  query: gql`
+    query listOrganizationReadOnlyGroups(
+      $filter: GroupFilter
+      $organizationUri: String!
+    ) {
+      listOrganizationReadOnlyGroups(
+        organizationUri: $organizationUri
+        filter: $filter
+      ) {
+        count
+        page
+        pages
+        hasNext
+        hasPrevious
+        nodes {
+          groupUri
+        }
+      }
+    }
+  `
+});
